+++ src/markitdown/converters/__init__.py
@@ -5,10 +5,16 @@
 from ._base import DocumentConverter, DocumentConverterResult
 from ._plain_text_converter import PlainTextConverter
 from ._html_converter import HtmlConverter
+from ._rss_converter import RssConverter
+from ._wikipedia_converter import WikipediaConverter
+from ._youtube_converter import YouTubeConverter
 
 __all__ = [
     "DocumentConverter",
     "DocumentConverterResult",
     "PlainTextConverter",
     "HtmlConverter",
+    "RssConverter",
+    "WikipediaConverter",
+    "YouTubeConverter",
 ]
